Moved all test types to the model package
