+++ types-jdbi/src/test/java/dk/cloudcreate/essentials/types/jdbi/model/CustomerIdArgumentFactory.java
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-package dk.cloudcreate.essentials.types.jdbi.types;
+package dk.cloudcreate.essentials.types.jdbi.model;
 
-import dk.cloudcreate.essentials.types.jdbi.LongTypeArgumentFactory;
+import dk.cloudcreate.essentials.types.jdbi.CharSequenceTypeArgumentFactory;
 
-public class OrderIdArgumentFactory extends LongTypeArgumentFactory<OrderId> {
-
+public class CustomerIdArgumentFactory extends CharSequenceTypeArgumentFactory<CustomerId> {
 }
