+++ types-jdbi/src/test/java/dk/cloudcreate/essentials/types/jdbi/model/OrderId.java
@@ -14,27 +14,28 @@
  * limitations under the License.
  */
 
-package dk.cloudcreate.essentials.types.jdbi.types;
+package dk.cloudcreate.essentials.types.jdbi.model;
 
 import dk.cloudcreate.essentials.types.*;
 
-import java.util.UUID;
+import java.util.Random;
 
-public class ProductId extends CharSequenceType<ProductId> implements Identifier {
-    public ProductId(CharSequence value) {
+public class OrderId extends LongType<OrderId> implements Identifier {
+    private static Random RANDOM_ID_GENERATOR = new Random();
+
+    public OrderId(Long value) {
         super(value);
     }
 
-    public static ProductId of(CharSequence value) {
-        return new ProductId(value);
+    public static OrderId of(long value) {
+        return new OrderId(value);
     }
 
-    public static ProductId ofNullable(CharSequence value) {
-        return value != null ? new ProductId(value) : null;
+    public static OrderId ofNullable(Long value) {
+        return value != null ? new OrderId(value) : null;
     }
 
-    public static ProductId random() {
-        // You can use any random Id generator like e.g. https://github.com/codahale/time-id
-        return new ProductId(UUID.randomUUID().toString());
+    public static OrderId random() {
+        return new OrderId(Math.abs(RANDOM_ID_GENERATOR.nextLong()));
     }
 }
